+++ requirements.txt
@@ -1,3 +1,4 @@
 pytest==6.2.2
 requests==2.25.0
 jsonschema==3.2.0
+PyYAML==5.4.1
