+++ test/support/schemas/get_user_info_schema.yml
@@ -1,0 +1,36 @@
+{
+    "type": "object",
+    "properties": {
+        "country": {
+            "type": "string"
+        },
+        "display_name": {
+            "type": "string"
+        },
+        "email": {
+            "type": "string"
+        },
+        "followers": {
+            "type": "object"
+        },
+        "href": {
+            "type": "string"
+        },
+        "id": {
+            "type": "string"
+        },
+        "type": {
+            "type": "string"
+        },
+        "uri": {
+            "type": "string"
+        }
+    },
+    "additionalProperties": true,
+    "required": [
+        "display_name",
+        "email",
+        "id",
+        "uri"
+    ]
+}
