+++ test_cases/API_table.md
@@ -14,7 +14,7 @@
 |TC-2|✔|1. there is an existing album with id: *{album_id}*<br>2. the *{OAUTH_TOKEN}* is added to the GET request's header<br>3. the *{album_id}* is added to the GET request|1. the "GET album by id" request is sent|1. the response should contain the proper album label
 |TC-3|✔|1. there is an existing album with id: *{album_id}*<br>2. the *{OAUTH_TOKEN}* is added to the GET request's header<br>3. the *{album_id}* is added to the GET request|1. the "GET album by id" request is sent|1. the response should contain the proper album name
 |TC-4|✔|1. the *{OAUTH_TOKEN}* is added to the GET request's header|1. the "GET current user profile" request is sent|1. the response should be successfully validated against the corresponding schema defined in folder: schemas
-|TC-5|✔|1. the *{OAUTH_TOKEN}* is added to the POST request's header<br>2. the *{playlist_name}* is added to the POST request's body<br>3. the *{playlist_desc}* is added to the POST request's body<br>4. the *{playlist_is_public}* is added to the POST request's body|1. the "POST create playlist" request is sent|1. the response should contain status code 201
+|TC-5|✔|1. the *{OAUTH_TOKEN}* is added to the POST request's header<br>2. the *{playlist_name}* is added to the POST request's body<br>3. the *{playlist_desc}* is added to the POST request's body<br>4. the *{playlist_is_public}* is added to the POST request's body|1. the "POST create playlist" request is sent|1. the response should contain status code *201*
 |TC-6|✔|1. the *{OAUTH_TOKEN}* is added to the POST request's header<br>2. the POST request's body does not contain the required field *{playlist_name}*|1. the "POST create playlist" request is sent|1. the response should contain status code *400*
 |TC-7|✔|1. the *{OAUTH_TOKEN}* is added to the POST request's header<br>2. the *{playlist_name}* is added to the POST request's body<br>3. the *{playlist_desc}* is added to the POST request's body<br>4. the *{playlist_is_public}* is added to the POST request's body|1. the "POST create playlist" request is sent|1. the response should contain the proper playlist *name, label* and *visibility*
 |TC-8|✔|1. the *{OAUTH_TOKEN}* is added to the POST request's header<br>2. the *{playlist_name}* is added to the POST request's body<br>3. the *{track_uris}* is added to the POST request's body|1. the "POST add items to playlist" request is sent|1. the response should contain status code *201*
@@ -28,7 +28,7 @@
 |TC-16|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. the playlist is empty<br>3. the *{OAUTH_TOKEN}* is added to the DELETE request's header<br>4. the *{playlist_id}* is added to the DELETE request<br>5. the *{tracks}* are added to the request's body|1. the "DELETE remove items from a playlist" request is sent|1. the response should contain status code *200*
 |TC-17|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. the *{bearer_token: ""}* is added to the GET request's header|1. the "GET playlist by id" request is sent|1. the response should contain status code *400*
 |TC-18|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. the *{bearer_token: "SOMERANDOMStr123"}* is added to the GET request's header|1. the "GET playlist by id" request is sent|1. the response should contain status code *401*
-|TC-19|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. an expired token: *{bearer_token: {EXPIRED_TOKEN}* is added to the GET request's header|1. the "GET playlist by id" request is sent|1. the response should contain status code *400*
+|TC-19|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. an expired token: *{bearer_token: {EXPIRED_TOKEN}}* is added to the GET request's header|1. the "GET playlist by id" request is sent|1. the response should contain status code *400*
 |TC-20|✔|1. there is an existing playlist with id: *{playlist_id}*<br>2. the *{bearer_token: None}* is added to the GET request's header|1. the "GET playlist by id" request is sent|1. the response should contain status code *401*
 |TC-21|❌|1. the Edge browser is opened|1. we navigate to "https://open.spotify.com/"|1. the Spotify web player should be opened
 |TC-22|❌|1. the Internet Explorer 1.0 browser is opened|1. we navigate to "https://open.spotify.com/"|1. the Spotify web player should send the following error message: "Browser not supported"
